Made building a lot easier, added missing full-text of CC0, added rmFile function
Build details as follows:
Just run a script with all required parameters and make sure to have a signing key on-hand.
If on Windows, revert to the previous plan where IDEA is used for building.
(It's not like it isn't standard over there to use the IDE as the main build platform.)